Move `get_to_thread()` to `scalems.utility`.
Requires moving `scalems.wait()` support to `scalems.workflow` to avoid
cyclic dependencies. This is appropriate, because `scalems.utility`
should not depend on other scalems submodules.

Supports #96.
