+++ src/scalems/utility.py
@@ -4,15 +4,17 @@ __all__ = [
     'app',
     'command',
     'function_wrapper',
+    'get_to_thread',
     'make_parser',
+    'next_monotonic_integer',
     'parser',
     'poll',
-    'wait',
     'ScriptEntryPoint'
 ]
 
 import abc
 import argparse
+import asyncio
 import contextvars
 import functools
 import logging
@@ -20,8 +22,6 @@ import typing
 from typing import Protocol
 
 from scalems import exceptions as _exceptions
-from . import get_scope
-from .workflow import WorkflowManager
 
 logger = logging.getLogger(__name__)
 logger.debug('Importing {}'.format(__name__))
@@ -219,135 +219,6 @@ def poll():
     raise _exceptions.MissingImplementationError()
 
 
-ResultType = typing.TypeVar('ResultType')
-
-
-class WorkflowObject(typing.Generic[ResultType]):
-    ...
-
-
-def _unpack_work(ref: dict):
-    """Temporary handler for ad hoc dict-based input.
-
-    Unpack and serialize the nested task descriptions.
-
-    Note: this assumes work is nested, with only one item per "layer".
-    """
-    assert isinstance(ref, dict)
-    implementation_identifier = ref.get('implementation', None)
-    message: dict = ref.get('message', None)
-    if not isinstance(implementation_identifier, list) or not isinstance(message, dict):
-        raise _exceptions.DispatchError('Bug: bad schema checking?')
-
-    command = implementation_identifier[-1]
-    logger.debug(f'Unpacking a {command}')
-    # Temporary hack for ad hoc schema.
-    if command == 'Executable':
-        # generate Subprocess
-        from scalems.subprocess import SubprocessInput, Subprocess
-        input_node, task_node, output_node = message['Executable']
-        kwargs = {
-            'argv': input_node['data']['argv'],
-            'stdin': input_node['data']['stdin'],
-            'stdout': task_node['data']['stdout'],
-            'stderr': task_node['data']['stderr'],
-            'environment': input_node['data']['environment'],
-            'resources': task_node['input']['resources']
-        }
-        bound_input = SubprocessInput(**kwargs)
-        item = Subprocess(input=bound_input)
-        yield item
-        return item.uid()
-    else:
-        # If record bundles dependencies, identify them and yield them first.
-        try:
-            depends = ref['message'][command]['input']
-        except AttributeError:
-            depends = None
-        if depends is not None:
-            logger.debug(f'Recursively unpacking {depends}')
-            dependency: typing.Optional[bytes] = yield from _unpack_work(depends)
-        else:
-            dependency = None
-        if 'uid' not in ref:
-            ref['uid'] = next_monotonic_integer().to_bytes(32, 'big')
-        uid: bytes = ref['uid']
-        if dependency is not None:
-            logger.debug('Replacing explicit input in {} with reference: {}'.format(
-                uid.hex(),
-                dependency.hex()
-            ))
-            ref['message'][command]['input'] = dependency
-        # Then yield the dependent item.
-        yield ref
-        return uid
-
-
-@functools.singledispatch
-def _wait(ref, *, manager):
-    """Use the indicated workflow manager to resolve a reference to a workflow item."""
-    raise _exceptions.DispatchError('No dispatcher for this type of reference.')
-    # TODO: Return an object supporting the result type interface.
-
-
-@_wait.register
-def _(ref: dict, *, manager):
-    # First draft: monolithic implementation directs the workflow manager to add tasks and execute them.
-    # TODO: Use a WorkflowManager interface from the core data model.
-    if not isinstance(manager, WorkflowManager):
-        raise _exceptions.ProtocolError('Provided manager does not implement the required interface.')
-    for item in _unpack_work(ref):
-        view = manager.add_item(item)
-        logger.debug('Added {}: {}'.format(
-            view.uid().hex(),
-            str(item)))
-    # TODO: If dispatcher is running, wait for the results.
-    # TODO: If dispatcher is not running, can we trigger it?
-
-
-# def wait(ref: WorkflowObject[ResultType], **kwargs) -> ResultType:
-def wait(ref):
-    """Resolve a workflow reference to a local object.
-
-    *wait* signals to the SCALE-MS framework that it is time to intervene and
-    do some workflow execution management.
-
-    ScaleMS commands return abstract references to work without waiting for the
-    work to execute. Other ScaleMS commands can operate on these references,
-    relying on the framework to manage data flow.
-
-    If you need to extract a concrete result, or otherwise force data flow resolution
-    (blocking the current code until execution and data transfer are complete),
-    you may use scalems.wait(ref) to convert a workflow reference to a concrete
-    local result.
-
-    Note that scalems.wait() can allow the current scope to yield to other tasks.
-    Developers should use scalems.wait() instead of native concurrency primitives
-    when coding for dynamic data flow.
-    However, the initial implementation does not inspect the context to allow
-    such context-sensitive behavior.
-
-    .. todo:: Establish stable API/CPI for tasks that create other tasks or modify the data flow graph during execution.
-
-    scalems.wait() will produce an error if you have not configured and launched
-    an execution manager in the current scope.
-
-    .. todo:: Acquire asyncio event loop from WorkflowManager.
-        scalems.wait is primarily intended as an abstraction from
-        https://docs.python.org/3.8/library/asyncio-eventloop.html#asyncio.loop.run_until_complete
-        and an alternative to `await`.
-    """
-    context = get_scope()
-    if context is None:
-        # Bail out.
-        raise _exceptions.DispatchError(str(ref))
-    if not isinstance(context, WorkflowManager):
-        raise _exceptions.ProtocolError('Expected WorkflowManager. Got {}'.format(repr(context)))
-
-    # Dispatch on reference type.
-    return _wait(ref, manager=context)
-
-
 def deprecated(explanation: str):
     """Mark a deprecated definition.
 
@@ -402,3 +273,35 @@ def next_monotonic_integer() -> int:
 
 
 _monotonic_integer = contextvars.ContextVar('_monotonic_integer', default=0)
+_T = typing.TypeVar('_T')
+
+
+@cache
+def get_to_thread() \
+        -> typing.Callable[
+            [
+                typing.Callable[..., _T],
+                typing.Tuple[typing.Any, ...],
+                typing.Dict[str, typing.Any]
+            ],
+            typing.Coroutine[typing.Any, typing.Any, _T]]:
+    """Provide a to_thread function.
+
+    asyncio.to_thread() appears in Python 3.9, but we only require 3.8 as of this writing.
+    """
+    try:
+        from asyncio import to_thread as _to_thread
+    except ImportError:
+        async def _to_thread(__func: typing.Callable[..., _T],
+                             *args: typing.Any,
+                             **kwargs: typing.Any) -> _T:
+            """Mock Python to_thread for Py 3.8."""
+            wrapped_function: typing.Callable[[], _T] = functools.partial(__func, *args,
+                                                                          **kwargs)
+            assert callable(wrapped_function)
+            loop = asyncio.get_event_loop()
+            coro: typing.Awaitable[_T] = loop.run_in_executor(
+                None, wrapped_function)
+            result = await coro
+            return result
+    return _to_thread
